Change package name to 'main'
The package name in the 'repeat.go' file in the '02-iteration' directory was changed from 'iteration' to 'main'. This modification will alter how the file is referenced in the rest of the Go code files. The change was made to make the file executable as the main package in Go.